+++ web/geo/generateAggregates.ts
