+++ web/geo/validate.ts
@@ -13,7 +13,7 @@ import {
 } from '@turf/turf';
 import { getHoles, getPolygons, log, writeJSON } from './utilities';
 
-import { mergeZones } from '../scripts/generate-zones-config';
+import { mergeZones } from '../scripts/generateZonesConfig';
 
 // TODO: Improve this function so each check returns error messages,
 // so we can show all errors instead of taking them one at a time.
