+++ web/scripts/generateZonesConfig.ts
